Create ActiveDirectory_FQDN_IP
Script for extracting all of the machine FQDNs from an Active Directory environment and exporting to a text file. Using that text files also perform an NSlookup to identify the last associated IP address of the machine and produce a clean list of IP addresses to be fed into another script, psexec, etc.