Tool for collecting and comparing file hashes.
To be used for the automated process of generating and comparing file hashes for copied files. Feed it origin path, copied path, and recursive or not. The script will hash and compare the files. The tool will also generate reports for both sets of files and any files with differences.